+++ src/main/java/roomescape/dto/internal/ReservationDetail.java
@@ -1,0 +1,11 @@
+package roomescape.dto.internal;
+
+import roomescape.entity.Payment;
+import roomescape.entity.Reservation;
+
+public record ReservationDetail(
+        Reservation reservation,
+        Payment payment,
+        long rank
+) {
+}
