+++ src/main/java/roomescape/dto/internal/ReservationWithRank.java
@@ -1,0 +1,9 @@
+package roomescape.dto.internal;
+
+import roomescape.entity.Reservation;
+
+public record ReservationWithRank(
+        Reservation reservation,
+        long rank
+) {
+}
